tests/wpt/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-004.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#baseline-rules">
<link rel="help" href="https://drafts.csswg.org/css-flexbox-1/#cross-alignment">
<style>
#target {
  display: flex;
  flex-wrap: wrap-reverse;
  flex-direction: column;
  align-items: last baseline;
  width: 200px;
  border: solid 3px;
  position: relative;
}
#target > :nth-child(1) {
  background: lime;
  margin-right: 20px;
  padding-right: 20px;
  font-size: 20px;
  line-height: 20px;
  align-self: first baseline;
  writing-mode: vertical-rl;
}
#target > :nth-child(2) {
  background: hotpink;
  font-size: 30px;
  line-height: 30px;
  writing-mode: vertical-lr;
}
#target > :nth-child(3) {
  background: papayawhip;
  font-size: 16px;
  line-height: 16px;
  writing-mode: vertical-lr;
}
#target > :nth-child(4) {
  background: orange;
  margin-left: 20px;
  padding-left: 20px;
  font-size: 20px;
  line-height: 20px;
  align-self: first baseline;
  writing-mode: vertical-lr;
}
#target > :nth-child(5) {
  background: cyan;
  font-size: 30px;
  line-height: 30px;
  writing-mode: vertical-rl;
}
#target > :nth-child(6) {
  background: papayawhip;
  font-size: 16px;
  line-height: 16px;
  writing-mode: vertical-rl;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('#target > div')">

<div id="target">
  <div data-offset-x="15">line1<br>line2</div>
  <div data-offset-x="0">line1<br>line2</div>
  <div data-offset-x="21">line1<br>line2</div>
  <div data-offset-x="125">line1<br>line2</div>
  <div data-offset-x="140">line1<br>line2</div>
  <div data-offset-x="147">line1<br>line2</div>
</div>
