LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-004-expected.txt

line1
line2
line1
line2
line1
line2
line1
line2
line1
line2
line1
line2

PASS #target > div 1
PASS #target > div 2
FAIL #target > div 3 assert_equals:
<div data-offset-x="21">line1<br>line2</div>
offsetLeft expected 21 but got 20
PASS #target > div 4
PASS #target > div 5
PASS #target > div 6
